requirements.txt
feedparser
google-generativeai
jinja2
pyyaml
python-dotenv
markdown
